silence linux warning

+++ RetroFE/Source/Graphics/Component/Image.h
@@ -96,7 +96,7 @@ private:
     std::vector<SharedSurface> animatedSurfaces_;
     std::vector<int> frameDelays_;
 
-    int currentFrame_ = 0;
+    size_t currentFrame_ = 0;
     Uint32 animationStartTime_ = 0;
     size_t lastRenderedFrame_ = std::numeric_limits<size_t>::max();
